# resources/openapi/checkout.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ecommerce Checkout API",
    "version": "0.1.0",
    "summary": "The boundary between a basket and a commitment, over the wire — with the domain's idempotency contract carried onto HTTP rather than reinvented on top of it.",
...
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "contact": {
      "name": "Liberu Software",
      "url": "https://liberusoftware.com"
    }
  },
  "servers": [
    {
      "url": "/{prefix}/{version}",
      "description": "Wherever the host mounted the groups. Both segments are configuration.",
      "variables": {
        "prefix": {
          "default": "api",
          "description": "`checkout-api.prefix`."
        },
        "version": {
          "default": "v1",
          "enum": [
            "v1"
          ],
          "description": "`checkout-api.version`."
        }
      }
    }
  ],
  "security": [
    {
      "checkoutToken": []
    }
  ],
...
  "paths": {
...
    "/checkout/sessions/{token}/placement": {
      "post": {
        "operationId": "placeCheckout",
        "summary": "Commit the checkout, exactly once",
        "description": "The one operation in this document a retry must not duplicate, and the reason `Idempotency-Key` is **required** here.\n\nIt does not create an order. It decides that the session is open, unexpired, has lines, has somebody to tell, and that the `authorized` plus `captured` tenders sum **exactly** to the grand total — and then emits `CheckoutCompleted`, which is what an order module subscribes to.\n\n## The key\n\nCaller-supplied, at most 191 characters, scoped `('checkout.place', key)` by the domain. It identifies *this attempt to place this checkout*: mint one when you render the Place Order button, keep it across every retry, and **do not regenerate it on each click** — a key that changes per click protects nothing.\n\nIt is **not** optional and it is not minted server-side, because a server-minted key would be a new key on every retry, which is the same as no key at all. A request without one answers **400**: a fault in the request itself that no change to the body can fix.\n\n## The four outcomes\n\n| Situation | Status | `Idempotency-Replayed` |\n| --- | --- | --- |\n| New key | `200` | `false` — the commit happened here, and `CheckoutCompleted` was dispatched |\n| Same key, same payload | `200` | `true` — the stored result is replayed, no work runs, **no event is dispatched** |\n| Same key, different payload | `409` | — the key was spent on a different checkout |\n| Key claimed, work unfinished | `423` | — retry unchanged, after `Retry-After` |\n\n**Both successes are 200 on purpose.** A client must not have to branch on a status code to find out whether its order exists; the flag is in the `Idempotency-Replayed` header and in the `replayed` field of the body, and the `data` object is identical either way.\n\n**409 and 423 are different answers to different questions.** 409 is permanent: two genuinely different requests shared a key, and the only safe answers are \"conflict\" or \"commit twice\" — replaying the first result would be the third and worst, a success returned for something that never happened. 423 is transient: a first call has claimed the key and has not finished, so the honest answer is \"ask again in a moment\", and answering 409 there would tell a client to give up on a request that is at that instant succeeding. It is never an empty success.\n\n## A failed attempt does not burn the key\n\nThe claim is released when the work throws, so a placement refused by its guards — no email yet, tenders short of the total, session expired — leaves the key usable. Fix what was wrong and retry with **the same key**, which is what a client naturally does. That is why 422 here is not the end of a key's life.\n\nThe idempotency record is kept for `checkout.idempotency.retention_days` (30 by default). **That window must exceed the longest retry window any of your clients uses** — a key pruned while a client still holds it stops being idempotent, and the next retry places a second order.",
        "tags": [
          "shopper"
        ],
        "deprecated": false,
        "x-liberu-deprecation": {
          "since": null,
          "sunset": null,
          "successor": null
        },
        "x-liberu-route-group": "shopper",
        "x-liberu-anonymous": true,
        "security": [
          {
            "checkoutToken": [
              "checkout:shopper.write"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/token"
          },
          {
            "$ref": "#/components/parameters/idempotencyKey"
          }
        ],
        "responses": {
          "200": {
            "description": "The placed checkout. `replayed` says whether this call did the work or replayed an earlier one; `data` is identical either way.",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "X-RateLimit-Limit": {
                "$ref": "#/components/headers/XRateLimitLimit"
              },
              "X-RateLimit-Remaining": {
                "$ref": "#/components/headers/XRateLimitRemaining"
              },
              "Idempotency-Replayed": {
                "$ref": "#/components/headers/IdempotencyReplayed"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlacementEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "423": {
            "$ref": "#/components/responses/Locked"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "checkoutToken": {
        "type": "oauth2",
        "description": "Scopes are enforced against whatever the actor answers to `tokenCan()` — Sanctum's abilities, Passport's scopes, or your own. **An actor with no such method carries no scope list to check** and is governed by the policy alone; that is a deliberate limit rather than a gap, because inventing a scope list for an actor that has none would mean denying every session-authenticated caller or admitting every one, and both are lies about what was checked.\n\nThe shopper group usually has no actor at all, and then there is nothing to narrow: what the request may reach is decided by possession of a token that names one checkout.\n\nThe audience is what is scoped, not the resource — driving one shopper's checkout and listing every open session in a team are categorically different grants.",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://your-host/oauth/token",
            "scopes": {
              "checkout:shopper.read": "Read a checkout session the caller holds the token for.",
              "checkout:shopper.write": "Drive a checkout session: contact, discount, consent, tender, placement, abandonment.",
              "checkout:staff.read": "Read checkout sessions in the actor's own team."
            }
          }
        }
      }
    },
    "parameters": {
      "token": {
        "name": "token",
        "in": "path",
        "required": true,
        "description": "The session's opaque token. A bearer credential — 48 alphanumeric characters, and the only handle a guest has on their own checkout.",
        "schema": {
          "type": "string",
          "pattern": "^[A-Za-z0-9]{20,64}$"
        }
      },
      "idempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Names *this attempt to place this checkout*. Mint one when you render the Place Order button and keep it across every retry; a key regenerated per click protects nothing. Required — a server-minted key would be a new key on every retry, which is the same as no key at all.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 191
        }
      },
...
    },
    "headers": {
      "XRequestId": {
        "description": "The caller's `X-Request-Id` when they sent one, or a minted uuid. Quote it when reporting a failure.",
        "schema": {
          "type": "string"
        }
      },
      "XRateLimitLimit": {
        "description": "Requests permitted per minute.",
        "schema": {
          "type": "integer"
        }
      },
      "XRateLimitRemaining": {
        "description": "Requests left in the window.",
        "schema": {
          "type": "integer"
        }
      },
      "IdempotencyReplayed": {
        "description": "`true` when this response replays an earlier commit made under the same key — in which case no work ran and no event was dispatched. `false` when the commit happened on this call.",
        "schema": {
          "type": "string",
          "enum": [
            "true",
            "false"
          ]
        }
      }
    },
    "schemas": {
      "Money": {
        "type": "object",
        "description": "An amount of money. `minor` is the authority — integer minor units, no float anywhere in this fleet — and `exponent` travels with it because a consumer handed `1999` cannot render it without knowing where the point goes. `decimal` is a **string**, because a JSON number is a float in most parsers and `19.99` does not exist as one.",
        "required": [
          "minor",
          "currency",
          "exponent",
          "decimal"
        ],
        "additionalProperties": false,
        "properties": {
          "minor": {
            "type": "integer",
            "examples": [
              1999
            ]
          },
          "currency": {
            "type": "string",
            "minLength": 3,
            "maxLength": 3,
            "examples": [
              "GBP"
            ]
          },
          "exponent": {
            "type": "integer",
            "minimum": 0,
            "examples": [
              2
            ]
          },
          "decimal": {
            "type": "string",
            "examples": [
              "19.99"
            ]
          }
        }
      },
      "Totals": {
        "type": "object",
        "description": "`net = subtotal - discount` and `grand_total = net + tax`, by construction on every session.",
        "required": [
          "subtotal",
          "discount",
          "net",
          "tax",
          "grand_total"
        ],
        "properties": {
          "subtotal": {
            "$ref": "#/components/schemas/Money"
          },
          "discount": {
            "$ref": "#/components/schemas/Money"
          },
          "net": {
            "$ref": "#/components/schemas/Money"
          },
          "tax": {
            "$ref": "#/components/schemas/Money"
          },
          "grand_total": {
            "$ref": "#/components/schemas/Money"
          }
        }
      },
      "Line": {
        "type": "object",
        "description": "One frozen line. `name`, `sku` and the prices are **copies** — a line must keep meaning something after the product it names has been renamed, re-priced or deleted — and `product_id` and `variant_id` are plain numbers with no relation behind them.",
        "required": [
          "kind",
          "name",
          "quantity",
          "taxable",
          "unit_price",
          "subtotal",
          "discount",
          "net",
          "tax",
          "gross"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "item",
              "shipping",
              "fee"
            ],
            "description": "Shipping is a line rather than a column, so the discount rule, the tax rule and the rounding have exactly one implementation."
          },
          "name": {
            "type": "string"
          },
          "sku": {
            "type": [
              "string",
              "null"
            ]
          },
          "quantity": {
            "type": "integer",
            "minimum": 1
          },
          "product_id": {
            "type": [
              "integer",
              "null"
            ]
          },
          "variant_id": {
            "type": [
              "integer",
              "null"
            ]
          },
          "taxable": {
            "type": "boolean"
          },
          "tax_rate_bp": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Basis points. Null when the tax arrived as an amount, which is what makes a later discount unreallocatable."
          },
          "position": {
            "type": "integer"
          },
          "unit_price": {
            "$ref": "#/components/schemas/Money"
          },
          "subtotal": {
            "$ref": "#/components/schemas/Money"
          },
          "discount": {
            "$ref": "#/components/schemas/Money"
          },
          "net": {
            "$ref": "#/components/schemas/Money"
          },
          "tax": {
            "$ref": "#/components/schemas/Money"
          },
          "gross": {
            "$ref": "#/components/schemas/Money"
          },
          "metadata": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": true
          }
        }
      },
      "Tender": {
        "type": "object",
        "description": "Provider-neutral. `provider` is an opaque string the host chose and this API never interprets.",
        "required": [
          "kind",
          "status",
          "amount"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "payment",
              "gift_card",
              "store_credit"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "authorized",
              "captured",
              "failed",
              "voided"
            ],
            "description": "Only `authorized` and `captured` count towards the amount owed."
          },
          "provider": {
            "type": [
              "string",
              "null"
            ]
          },
          "reference": {
            "type": [
              "string",
              "null"
            ]
          },
          "amount": {
            "$ref": "#/components/schemas/Money"
          }
        }
      },
      "Consent": {
        "type": "object",
        "description": "Evidence, redacted for the wire. The IP address and the user agent captured with the agreement are **deliberately absent** from every response in this document.",
        "required": [
          "type",
          "document_version",
          "agreed",
          "agreed_at"
        ],
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "examples": [
              "terms"
            ]
          },
          "document_version": {
            "type": "string",
            "examples": [
              "2026-01-14"
            ]
          },
          "document_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "agreed": {
            "type": "boolean",
            "description": "False is a first-class value: a shopper who declined has said something."
          },
          "agreed_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PlacedCheckout": {
        "type": "object",
        "description": "The frozen record of a commit. A different shape from a session rather than a session with `status: placed`: there is no `expires_at` left to watch and no discount left to change, and the two fields it gains are the two a retry needs to recognise its own receipt.",
        "required": [
          "token",
          "currency",
          "exponent",
          "totals",
          "lines",
          "tenders",
          "consents",
          "idempotency_key",
          "placed_at"
        ],
        "properties": {
          "token": {
            "type": "string"
          },
          "currency": {
            "type": "string",
            "minLength": 3,
            "maxLength": 3
          },
          "exponent": {
            "type": "integer",
            "minimum": 0
          },
          "email": {
            "type": [
              "string",
              "null"
            ],
            "format": "email"
          },
          "customer_id": {
            "type": [
              "integer",
              "null"
            ]
          },
          "shipping_address": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": true
          },
          "billing_address": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": true
          },
          "totals": {
            "$ref": "#/components/schemas/Totals"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Line"
            }
          },
          "tenders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tender"
            }
          },
          "consents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Consent"
            }
          },
          "idempotency_key": {
            "type": "string",
            "description": "The key this commit was made under. Also the correlation id an order listener should be idempotent on."
          },
          "placed_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PlacementEnvelope": {
        "type": "object",
        "required": [
          "data",
          "replayed"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PlacedCheckout"
          },
          "replayed": {
            "type": "boolean",
            "description": "True when this response replays an earlier commit made under the same key. The same fact as the `Idempotency-Replayed` header, in the body, because a client reading a stored response has the body and not the headers."
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "The request itself is unusable and no change to the body fixes it. On `POST .../placement` this is a missing, blank or over-long `Idempotency-Key`.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "No such session. Identical for a token that names nothing and a token that names somebody else's checkout.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Conflict": {
        "description": "The resource is in a state that refuses the request, permanently: a session that is no longer open, or an idempotency key already spent on a different payload. Retrying this unchanged will always fail.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Locked": {
        "description": "The idempotency key is claimed by a call that has not finished. Transient, and its own status rather than a 409 — a client told \"conflict\" for an in-flight commit stops retrying, and the response it was waiting for never arrives. Retry unchanged after `Retry-After`.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unprocessable": {
        "description": "Well-formed, and the rules say no. Validation, or one of the domain's refusals: nothing to place, nobody to tell, tenders that do not sum to the total, a currency the session is not in, a discount larger than the subtotal, or a tax amount that cannot be reallocated. A placement refused here does **not** burn its idempotency key.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "The limiter refused. Carries `Retry-After`. Note that a placement retry is meant to be cheap and safe — a limit that refuses it is a limit that hides the receipt.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}